backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
huggingface-hub
openai
pydantic
requests
faiss-cpu
numpy
python-multipart
